Rename authdb project to auth_db
